\name{factor.a}
\alias{factor.a}
\docType{data}
\title{
Table: Factor for variable control charts
}
\description{
A data frame containing the factor for variable control charts calculations.
}
\usage{data(factor.a)}
\format{
A data frame with factors (ex: A2, d2, D4 and so on) for size groups from 2 to 25.
}
\source{
Montgomery, D.C. (2005) \emph{Introduction to Statistical Quality Control}, 5th ed. New York: John Wiley & Sons, ISBN 0-471-65631-3
}
\examples{
data(factor.a)
}
\keyword{factor}
\keyword{X_chart}
