\name{XRSCC-package}
\alias{XRSCC-package}
\alias{XRSCC}
\docType{package}
\title{
Calculates and plots variable and attributes control charts
}
\description{
Calculates the control limits for each type of variable or attribute control chart, then using an iteration to get the true control limits
}
\details{
\tabular{ll}{
Package: \tab XRSCC\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-05-04\cr
License: \tab GPL\cr
}

}
\author{
Erick Marroquin\cr Maintainer: Erick Marroquin <ericksuhel@gmail.com>
}
\keyword{package}
\keyword{quality_control}
